test(runtime-pi): share the container e2e gate and bundle fixture
`provision-container.e2e.test.ts` owned the whole container-e2e preamble
privately: the opt-in gate (TEST_DOCKER/CI), the docker/daemon/image platform
probes that make #882 skip honestly instead of failing inside `docker run`, the
minimal agent `.afps-bundle` the mock `/workspace` route serves, and the
`docker logs` dump that makes a deadline failure diagnosable. A second
container e2e is about to need all four, and a copied gate is a gate that
drifts — one suite silently stops running while its warning still reads right.

Lift them into `test/helpers/container-e2e.ts` with no behaviour change. The
gate now takes a list of images (a suite needing the pi+sidecar pair must check
both) and its rebuild hint names `bun run docker:build:runtime`, the only
command that builds the pair with matching revision stamps.

`codexPlaceholderJwt()` moves here too: `pi-runner-transport.test.ts` hand-rolled
the unsigned JWT that carries `chatgpt_account_id` into pi-ai's header
assembly, and the container e2e needs the same value in `MODEL_API_KEY`. It is
hand-rolled rather than imported because `@appstrate/module-codex` is not a
dependency of `runtime-pi`.
